Move stuff to more natural places
